Soil Rebalancing (wip)
Currently digging leaves you with shitloads of soil and not-so-many ways of using it.
This definitely needs to be fixed.

As a first step, modifying adobe brick recipe to accept thread as a source of fiber
(some sources state that you don't even need any of it to make a brick, see TODO).

TODO
- consider an alternative recipe for adobe bricks without fiber/ with more soil
- add recipe to sift soil (using water & sieve) resulting in sand and clay
- maybe more, make up as I go