Filthy crafting, Take 1
Allow using filthy items in crafting with the relevant game option.
Definitely needs more thought but the whole issue seems like a big ball of clusterfuck.
As a further thought - explicitly allow filthy stuff to be recipe's components (not unlike ALLOW_ROTTEN flag).

For now, it's "all or nothing".
Oh, and it'll be sure as hell bugged af =)